# Clientory App/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scans:
    post:
      operationId: createScan
      tags: [scans]
      summary: Create a new AI visibility scan
      parameters:
        - name: x-email-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanBody"
      responses:
        "201":
          description: Scan created and started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "200":
          description: Cached scan returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CachedScan"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Entitlement check failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntitlementErrorResponse"
        "422":
          description: Input validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Email verification required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitResponse"
  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get scan results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDetail"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scans/{id}/stream:
    get:
      operationId: streamScanProgress
      tags: [scans]
      summary: Stream scan progress via SSE
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE stream of scan progress events
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    EntitlementErrorResponse:
      type: object
      properties:
        error:
          type: string
        entitlementCode:
          type: string
          enum:
            - FREE_REPORT_ALREADY_USED
            - SUBSCRIPTION_REQUIRED
            - BUSINESS_PLAN_LIMIT_REACHED
            - SUBSCRIPTION_INACTIVE
            - BILLING_PERIOD_ENDED
        periodEnd:
          type: string
          format: date-time
          nullable: true
        scansUsed:
          type: number
          nullable: true
        scansLimit:
          type: number
          nullable: true
      required:
        - error
        - entitlementCode
    RateLimitResponse:
      type: object
      properties:
        error:
          type: string
        retryAfter:
          type: number
      required:
        - error
...
    CreateScanBody:
      type: object
      properties:
        businessName:
          type: string
        businessType:
          type: string
        location:
          type: string
        description:
          type: string
        website:
          type: string
        businessId:
          type: string
      required:
        - businessName
        - businessType
        - location
        - businessId
    Scan:
      type: object
      properties:
        id:
          type: string
        businessName:
          type: string
        businessType:
          type: string
        location:
          type: string
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, generating_prompts, scanning, completed, failed]
        score:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        businessId:
          type: string
          nullable: true
        isFreeReport:
          type: boolean
      required:
        - id
        - businessName
        - businessType
        - location
        - status
        - createdAt
    CachedScan:
      type: object
      properties:
        id:
          type: string
        businessName:
          type: string
        businessType:
          type: string
        location:
          type: string
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, generating_prompts, scanning, completed, failed]
        score:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        cached:
          type: boolean
      required:
        - id
        - businessName
        - businessType
        - location
        - status
        - createdAt
        - cached
    ScanPrompt:
      type: object
      properties:
        id:
          type: string
        scanId:
          type: string
        prompt:
          type: string
        category:
          type: string
        audience:
          type: string
          enum: [individual, business]
          nullable: true
      required:
        - id
        - scanId
        - prompt
        - category
    ScanResult:
      type: object
      properties:
        id:
          type: string
        scanPromptId:
          type: string
        provider:
          type: string
          enum: [openai, anthropic, gemini]
        response:
          type: string
        mentioned:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - scanPromptId
        - provider
        - response
        - mentioned
        - createdAt
    ScanDetail:
      type: object
      properties:
        scan:
          $ref: "#/components/schemas/Scan"
        prompts:
          type: array
          items:
            type: object
            properties:
              prompt:
                $ref: "#/components/schemas/ScanPrompt"
              results:
                type: array
                items:
                  $ref: "#/components/schemas/ScanResult"
            required:
              - prompt
              - results
        recommendations:
          type: array
          items:
            type: string
      required:
        - scan
        - prompts
        - recommendations
...